requirements.txt
fastapi
uvicorn[standard]
chromadb
pydantic
python-dotenv
google-genai
numpy
pypdf
python-multipart
starlette